The itemref-tooltip will nolonger disappear when holding down the shift-key. (fixes #122)
Split the OnHyperlink function into two seperate ones, so the EnhTT-part can be called indepenently. Had to do this, as the original TTfunction causes the bug. It's nolonger been called now, when redisplaying an old itemreftooltip.
Also moved all those globals to the top of the file, making it easier to work with those.
Unfortunately these changes result in the following:
- /auc embed on
- display a chatitemlink (don't close that window)
- /auc embed off
- move the mouse over a bagpack
- move the mouse off that backpack
Now both, the embeded and the external EnhTT is shown for the itemref item. The solution causing the bug above was designed to work around this problem, but it has to be "fixed" in another revision. I could not fix it myself, as before implementing this fix I'd like to discuss some changes to the script with norganna...

git-svn-id: http://svn.norganna.org/auctioneer/trunk@295 21a0ffd4-88d5-4261-b61c-461777ffcffb