Add temporary backwardly compatible functions to the new EnhTooltip code.
I have put some very simple backwardly compatible code at the bottom of the EnhTooltip to give dependant mod authors a chance to upgrade to the new tooltip format.
I will give them a month or so to upgrade, then we will start printing out messages that the author needs to upgrade soon into the chat window when the old functions are called.


git-svn-id: http://svn.norganna.org/auctioneer/trunk@308 21a0ffd4-88d5-4261-b61c-461777ffcffb