Fixed Khaos registration and bidbroker output
Several of the feedback entries for the Khaos configs were just strings instead of functions, I've wrapped them up proper.
The textbox to default a value by name now defaults to "" instead of nil, to prevent a Khaos error.  I've also increased it's difficulty to 4, developer.
Auctioneer_Default will now consider "" to be the same as nil, and will simply return if this is passed to it.  Otherwise on load of Khaos, all data would be cleared.  This means you must now pass all or _AUCT['CmdClearAll'] when using /default, which is probably safer anyway.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@331 21a0ffd4-88d5-4261-b61c-461777ffcffb