Added ability to prevent the auction house frame from getting hidden when opening other windows.
This is controlled by a function in AucUtil, Auctioneer_ProtectAuctionFrame, pass true to this function to protect the frame from being hidden when not desired, pass not true to set it normal.

This function is called from AucScanner line 394 as it needs to be called after the auction house loads.  This call should be changed to use an option.

Note: This will not work with world map frame, as it uses some funky internal code to force closing the AH.

Note 2: Some frames will still hide the auction frame, if the current version of MobileFrames is used, I will get AnduinLothar to fix this issue.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@483 21a0ffd4-88d5-4261-b61c-461777ffcffb