Changed the Auctioneer stub loader to be more reliable
Implemented a workaround for Stubby processing the AUCTION_HOUSE_SHOW before UIParent.
Auctioneer was assuming Blizzard_AuctionUI was loaded at this point, which was not always the case.
We now hook into functions instead of events to be able to guarantee the order in which some things are executed.

Also placed back the code Mugendai put in in revision [546]. It's only purpose is now to properly initialize AuctionFrame protection if Auctioneer is (manually) loaded while the AuctionFrame is already visible.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@548 21a0ffd4-88d5-4261-b61c-461777ffcffb