Fixed a problem with the Auctioneer Khaos registration. Users with this problem will either have to edit or delete their SavedVariables\Khaos.lua, however once edited\deleted the error will no longer appear.
The 
"Count: 1
Error: Interface\AddOns\Khaos\Khaos.lua:2342: Usage: SetText("text")"
error was generated because the Auctioneer_GetKhaosDefault() function was returning a boolean value instead of a string when called with the "also" paramenter. Since this setting is saved in the Khaos settings table the user will have to edit/delete the "["configuration"]["Auctioneer"]["also"]["value"]" entry themselves.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@593 21a0ffd4-88d5-4261-b61c-461777ffcffb