Added a BidManager to deal with the mechanics of searching for and bidding on items.
There is still some bullet proofing work to do. Primarily the search/bid ability should be disabled while processing bid requests. The basic mechanics should all be in place though. It would also be great if somehow we could tag items in the Search Auctions tab with the results of bid attempts. I'll have to do some thinking on that...

git-svn-id: http://svn.norganna.org/auctioneer/trunk@639 21a0ffd4-88d5-4261-b61c-461777ffcffb