Made some changes to the way the BidBroker (and filtering in general) work to support additional filtering in the new AuctionFrameSearch searching. Also add search-saving functionality.
The searching functions now allow a minimum quality parameter and a regex search on the name to be performed. Also the bidbroker has been altered to remove some of the limitations on minimum percentless and it also ignores profit if the stated minimum profit is 0 (which allows the user to put in negative percentless values and still get back valid results)
There is also a feature for saving and recalling search queries (say you like finding your cloth at certain prices etc)
The current functionality only works for bidding at the moment, I will be moving the functionality to buyouts and a new purely price-based search soon.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@652 21a0ffd4-88d5-4261-b61c-461777ffcffb