(fixes #1274)Fixes the bug concerning combining stacks of an item while disenchanting. (Putting file in the right place this time)
I traced the bug and found out that pickupContainerItemHook and pickupInventoryItemHook were still being called while the player was disenchanting. This caused the saved item (the item that enchantrix assumed was being disenchanted) to be changed to whatever was picked up. I added an if to both of those functions so that they are not called when the player is casting a spell, IE Disenchant. I tested this a few times, and it works for me, both when picking up and item and not. Feel free to test it more.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1283 21a0ffd4-88d5-4261-b61c-461777ffcffb