Fix AucFilter.lua line 126 error.  (fixes #1240)
Added check for non-nil itemInfo object.  I'm not sure why we are getting these, as I would think that they would have been caught and excluded prior to being put into the snapshot, but they are there, so we have to trap for them.  If we can figure out where they are coming from and fix them, we can maybe revert this check and gain a little speed on the bid/buyout search.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1309 21a0ffd4-88d5-4261-b61c-461777ffcffb