void
Moved setting the debug switch to the top of the file to easier find and adjust it. Also changed it from uncommenting to explicitly setting it to false. The old way (uncommenting, if not needed) could cause a global debug-variable/function from any other addon being used instead of the expected local one.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1337 21a0ffd4-88d5-4261-b61c-461777ffcffb