(fixes #1295) Stubby/Enhanced Tooltip could report the following error message:    ..\AddOns?\EnhTooltip\Tooltip.lua line 835: attempt to index local 'line' (a nil value) This has been fixed.
The function addSeparator() missed to check if the requested font string line exists. In case it does not, which occures whenever one tries to add a separator after the last line the tooltip, line is nil and will result in an errormessage on the next line.
The code was changed to make sure that a new line is being created for the separator, if it does not exist. Please also read the note on the new function getLine(), which explains a possible case in which the nil bug could still occure.

Additionally changed the code in createNewFontString() to make it more robust against failed function calls to one of blizzards UI functions, which might also cause the nil bug.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1342 21a0ffd4-88d5-4261-b61c-461777ffcffb