(fixes #1413) BTMScanner caused a nil-errormessage, if a thrid-party addon (in this case it was reported with GuildAds) passes an incorrect itemlink to Enhanced Tooltip. This has been fixed.
BtmScan.TooltipHook expects a correct itemlink, which might not be the case if called incorrectly. Since the function is globally accessable, since it's hooked into EnhTooltip.AddTooltip, any incorrect itemlink might cause an errormessage. Therefore a preliminary check has been added. Of course, now we loose the possibility to track down any own bug in this part of the code. Maybe some debugging system should be added in the future.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1395 21a0ffd4-88d5-4261-b61c-461777ffcffb