* removed an old workaround, since stubby's unhooking function works once again
Prior to the fix for stubbys unhooking function, auctioneer used a boolean to check if hookKill() should return "killorig" (while scanning) or not (if it was not scanning). Since the unhooking procedure works once again, this workaround is no longer needed and the proper functions can be used.
This reduces some overhead, since killHook won't be called when not scanning the ah, any longer.

In advance changed the way killHook is defined to the same way as all the other functions are.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1427 21a0ffd4-88d5-4261-b61c-461777ffcffb