Fix bug where canceling a bid or buyout would cause search/scan buttons to permanently disable.  (addresses #1304)
The static popup confirmation was not calling the bidCompleteCallback function if the user canceled the bid/BO, so the bid request was never removed from queue.  I am not yet closing #1304 because there are reports of similar problems with semi-permanently disabled search buttons when BTM is running which seem to have nothing to do with this canceled bid problem.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1428 21a0ffd4-88d5-4261-b61c-461777ffcffb