Fixes displaying the incorrect page number being scanned atm. Also updated the progressing message to be easier understandable.
There were a bunch of redundand bugs in the current UI display code which resulted in inappropriate messages being displayed. All of these have been fixed and in advance the UI processing function has been rewritten to use passed parameters instead of directly operating on the request "object".
The current way is only a temporary solution. A more permanent one might be introduced in 4.2.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1451 21a0ffd4-88d5-4261-b61c-461777ffcffb