A bug in stubby was fixed which could shift the return parameters of a hooked function and therefore cause other addons as well as some Auctioneer addons not to work as designed. This also fixes at least some, if not all of the symptoms reported in #1277.
Some Blizzard functions (in the case of #1277 it was GetInboxText() are designed to return nil values in between other return values (in the case of GetInboxText() it returned nil, "STATIONERYTEST", nil).
Stubby.hookCall() is designed to create a table, containing these return values, resulting in a table with holes. table.remove() was used to trim the first value from this created table to remove the pcall result code. Unfortunately table.remove() does not work with tables containing these holes as expected and in this case did not move the index of STATIONERYTEST. This resulted in the unpack() function unpacking and returning STATIONERYTEST as the third, instead of the second parameter.

I'm not absolutely sure what other symptoms are caused by this bug, but I doubt that the mailbox issue is the only one. Nevertheless, since there are no major tickets opened, I did not investigate, if that might have fixed any other tickets, as well.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1481 21a0ffd4-88d5-4261-b61c-461777ffcffb