code layout changes
Renamed the local self variable to EnhTTData, so it won't confuse us with the self variable, Blizzard uses, if a function is called in an xml declaration (see changes on WoW 2.0).
In advance we no longer hide blizzards variable by defining a local self variable which allows us access to blizzard functions in tooltip.lua.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1490 21a0ffd4-88d5-4261-b61c-461777ffcffb