(fixes #1346) Enhanced tooltip should work properly with any addon which sticks their tooltip to the cursor, now.
Refering to http://www.wowwiki.com/API_GameTooltip_SetOwner setting the offsets for x/y don't have any effect when the frame is anchored to the tooltip. Therefore I guess that the SetPoint() function is also not designed to work properly, if the anchor is set to the cursor and should not be used to move the frame.

Before this revision we did not check the case for ANCHOR_CURSOR and tried to move the current tooltip which resulted in this strange bug.

The current solution disables the relocation of the tooltip, if the current tooltip is anchored to the cursor. A better way would have been to code something to set the point, where the mousecursor is anchored, so that the complete tooltip will still be shown on screen. Nevertheless I hope that this fix will do fine for the moment. (a new ticket was opened, to keep this possible enhancement in mind for one of the next releases #1469)

Because quite alot of changes were necessary to properly fix this issue and I only tested the anchor code for a few minutes using CT_BuffMod, I'd appreciate any code reviewing and or further testing to make sure that no other bug made it into the code.

In advance added some more code comments.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1516 21a0ffd4-88d5-4261-b61c-461777ffcffb