Fixed the OnUpdate() method of beancounter being not executed.
The OnUpdate section of a frame is not executed for hidden frames. Therefore the complete OnUpdate()-code in MailMonitor.lua was never called at all. I'm not sure what issues arrose from this bug, but since nobody caught this bug since it's introduction in r821 I hardly doubt that any severe issues were caused by this.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1518 21a0ffd4-88d5-4261-b61c-461777ffcffb