Fixed a potential incompatibility with 3rd party addons which alter the parameterlist for CanSendAuctionQuery().
Removed the two optional parameters from CanSendAuctionQuery() and reverted the way hooking into the return value from the original CanSendAuctionQuery()-function is done to a modified version of the code before r1222.

Instead of passing parameters, a new function was added to BtmMain, which is equally used like that, defined in AucQueryManager.
Also the way AucQueryManager alters the CanSendAuctionQuery() was reverted to the way before r1222 meaning, that even if using AucQueryManager.CanSendAuctionQuery(), isQueryInProgress() and isBidInProgress() will be checked.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1520 21a0ffd4-88d5-4261-b61c-461777ffcffb