The remember price checkbutton in the auctions tab in the AH can no longer be checked/unchecked, if no item is selected in that frame. This makes it easier to understand/use that Auctioneer feature.
Partly fixes #1300, where the usage of the remember price feature was reported to be wrong or at least not very rational.
Before it was possible to uncheck that checkbox, if there was no item selected, but it was not possible to check it again. That could confuse users, who might wonder what this behaviour means.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1530 21a0ffd4-88d5-4261-b61c-461777ffcffb