improved performance for embeded tooltips
Improved handling and rendering of embeded tooltip lines especially for colored lines. Usage of AddLine() was improved to do displaying and coloring of a single line in one step instead of adding the line first and change its color afterwards. Also changed the way embedLines is being reset (nolonger needs to iterate over all entries) and is being iterated over (use the faster ipairs() function instead of the slower but unnecessary pairs() one).

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1628 21a0ffd4-88d5-4261-b61c-461777ffcffb