Corrected the description of the lib.OnLoad() function.
This function is called without arguments when the module itself is loaded, whereas lib.Processor("load") is called once for every single module with the name of the module as the second argument. Note that this means that the module will receive TWO load notifications if it both listens for lib.OnLoad() calls and lib.Processor("load") calls.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1635 21a0ffd4-88d5-4261-b61c-461777ffcffb