void
Added final version of debugLib. This is the intended version for the 4.2 and probably the 5.0 release.
Almost every part except the dump() function has been improved starting with the initialization code which now handles the absence of nLog better than before up to a big improvement to the parameter handling for DebugLib.DebugPrint().
A complete documentation was added, as well and the whole code was tested using different scenarios.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1723 21a0ffd4-88d5-4261-b61c-461777ffcffb