Fixes invalid item information being displayed/used. This should fix quite a lot of reported bugs in 4.0.1. For instance this fixes #1558.
The cache function introduced in r1669 in getItem() did not correctly differentiate between cached calls and non-cached calls. If a the function was called by in uncached mode, and the next call to this function was issued for the same item but in cached mode, the function returned an empty table instead of the proper item information.
Another scenario was that the function could return data for older items. This was caused by the same bug, but caused other symptoms for users.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1758 21a0ffd4-88d5-4261-b61c-461777ffcffb