void
Fixes another case which caused the tooltip becoming too wide (see #1443). The problem was that under some circumstances Blizzard tooltip was not updated before showTooltip() was called. Therefore Blizzard tooltip was still set to the tooltip's width of the previous tooltip and therefore we worked with the incorrect assumption that Blizzard's tooltip was wider than our own one.
EnhTT will now call currentTooltip:Show() before getting the tooltip's width.

In addition the unnecessary call to currentTooltip:Show() after embedRender() was removed, too, which increases the performance if all EnhTT information will be embedded in the currentTooltip.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1808 21a0ffd4-88d5-4261-b61c-461777ffcffb