Added support for "scanning" during a regular search. Added support for filtering a scan on any/all filter criteria.
This routine was primarily to allow the scanner to handle the data returned during normal search button activity.
It also commits data as soon as the Auction House Window is closed.
Following KISS, it has removed the Suspended state.  Items that are not seen are either deleted or ignored.
Items can be ignored only when an incomplete scan is being committed.



git-svn-id: http://svn.norganna.org/auctioneer/trunk@1811 21a0ffd4-88d5-4261-b61c-461777ffcffb