(fixes #1640, #1641) Enhanced tooltip will no longer cause a drastical fps dropdown while mousing over items in the mailbox. This also fixes the issue that enhTT won't be displayed for all new TBC items there.
The implementation for the SetInboxItem() hook iterated over all items (1 to 30000) to get the itemLink. Depending on what the real itemID was, that could last quite a while. While this has been present as of the first implementation in r855, it has not had that much of an impact over the performance, as most items had ids lower than 10000. With the new TBC items having IDs > 20000, that's now recognized as an annoying behavior by most users. A side effect of the old implementation was that items with an id > 30000 would not be recognized resulting in no enhanced tooltips being displayed.
The changed function now used GetInboxItemLink() rather than trying to iterate over all items.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1926 21a0ffd4-88d5-4261-b61c-461777ffcffb