(fixes #1647) Provide a backup set of item data for enchanting reagents
GetItemInfo fails if the item data is not in the user cache, so we got nil names and links for some enchanting reagents.
Also, we had an old cache of reagent data that wasn't actually used anywhere (now removed).
Now we install default data (English strings) for the reagents.
If the item is in the user's game cache, the default data will be overwritten by the user's localized data.
The new cache is only used as a backup in case GetItemInfo fails.



git-svn-id: http://svn.norganna.org/auctioneer/trunk@1953 21a0ffd4-88d5-4261-b61c-461777ffcffb