Improved speed of scanning and post-scanning processes.
Did some extensive profiling to see where the slow-down was with deleting auctions. Found that the nextID removal function was extremely expensive. Removed this list (which takes up a bunch of SV space also) with a once-per-session generated, sorted list, and a simple iterative getnextid function.
Found that the GetScanData was getting called once per item also, so added "begin" and 
"complete" events to the scanning process, where we can do once-off jobs, and clean up afterwards, respectively.
Also noticed that EnhTooltip's BreakLink function is quite expensive, and for the most part, unneeded for our primary DecodeLink function. Converted this function to use a strsplit/select based equivalent which is at least 20 times faster.
Scans are taking at most 2 seconds to process, post-scan now, instead of the 60+ seconds that they were taking before (at least on my computer that is).

git-svn-id: http://svn.norganna.org/auctioneer/trunk@1994 21a0ffd4-88d5-4261-b61c-461777ffcffb