Pulling Auctioneer and BeanCounter into the mainline trunk, where it will be forthwith built from.
This means that Auctioneer will from this point in time be built with the new build scripts and will utilize the new embeddable libs in place of svn externals.
Also it will use the new BtmScan and new everything else, and provides an alternative to the AuctioneerAdvanced in interface and personality only, marking the end of the Auctioneer 4.0 line, and the beginning of the AuctioneerClassic 5.0 line.


git-svn-id: http://svn.norganna.org/auctioneer/trunk@2283 21a0ffd4-88d5-4261-b61c-461777ffcffb