Add Protect Window functionality to AA (ADV-21)
Added a function, .windowProtect(), to CoreUtil.lua that is called by our control in CoreSettings.lua, and which protects the auction house window.  We modified the getter, setter, and getDefault functions in CoreSettings.lua to accept a function reference in the place of a setting name, which enables us to call my function in the control.  This means the function has to handle the getting and setting of the values from AucAdvanced.Settings via calls to AucAdvanced.Settings.(Get/Set)Attribute, but it also allows the status of the window to be updated when the setting is changed.  We would otherwise have needed to reload the UI to get changes to the settings to take effect.

It also adds in a default I failed to add when committing the print-in code, ADV 153

git-svn-id: http://svn.norganna.org/auctioneer/trunk@3020 21a0ffd4-88d5-4261-b61c-461777ffcffb