(CLAS-86) This marks the end of the long and fruitful life of the Auctioneer Classic code-base. It had its start with Auctioneer 2.0 and has been upgraded and modified ever since, sweeping the Auction House mod landscape in the process.
I will sincerely miss working on it, but its time to let it go into the annals of history. It simply won't handle the influx of items that WotLK will bring to the table.

I know that we originally told our users that we would support Auctioneer Classic for six months after Auctioneer Advanced's release, but this is simply not a practical option anymore. The amount of engineering required to refit it to work in WotLK is just too much for the delicate house of cards that is Auctioneer Classic's current state.

I hope you all understand as you always have,
--
MentalPower
Lead Developer of the Auctioneer AddOns project and Administrator of Norganna's AddOns

git-svn-id: http://svn.norganna.org/auctioneer/trunk@3572 21a0ffd4-88d5-4261-b61c-461777ffcffb