Implements [SDAT-6].  Stack size breakdown now always displayed.
The old code tried to check for there being only a single stack size listed in the snapshot.  If so, it would not display the stack size breakdown.  Unfortunately, the dist.stack table is sparse, and so the return of #(dist.stack) actually returns the size until the first non-existant stack (e.g. a 0 if there is no 1 stack, a 3 if stacks are (1,2,3,10,20), a 5 if they are (1,2,3,4,5,19), etc.).

Since there is no easy, clean way to count the table size for sparse tables, for the moment I've just removed the check altogether.  If saving that one line becomes important down the road, we can add it another table iteration to do a count first.

git-svn-id: http://svn.norganna.org/auctioneer/trunk@4411 21a0ffd4-88d5-4261-b61c-461777ffcffb